rearranged screen components
