+++ src/screens/NavBar.tsx
@@ -1,6 +1,6 @@
 import { Link, Outlet, useNavigate } from "react-router-dom";
 import React, { useState, useRef, useEffect, useCallback } from "react";
-import { useAuth } from "./App";
+import { useAuth } from "../App";
 
 const NavBar: React.FC = () => {
   //hook state for account accordion
@@ -23,7 +23,7 @@ const NavBar: React.FC = () => {
     }
   }, []);
 
-  //function to log out
+  //function to log out, will refactor this later
   const logout = () => {
     //remove all the tokens from browser
     localStorage.removeItem("access_token");
