+++ src/screens/VerifyFail.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import axiosPublicInstance from "./axiomPublic";
+import axiosPublicInstance from "../axiomPublic";
 
 const VerifyFail: React.FC = () => {
   const navigate = useNavigate();
